feat - add simulation client
Remove task for old sim command
