+++ src/inputs_outputs/mod.rs
@@ -1,3 +1,6 @@
-pub mod vision;
 pub mod game_controller;
 pub mod multicast_client;
+pub mod output;
+pub mod simulation_client;
+pub mod ubs_client;
+pub mod vision;
