+++ src/lib.rs
@@ -2,6 +2,7 @@
 #![warn(clippy::unwrap_used)]
 #![warn(clippy::panic)]
 // #![deny(clippy::all)]
-pub mod libs;
-pub mod inputs_outputs;
 pub mod filters;
+mod guard;
+pub mod inputs_outputs;
+pub mod libs;
